Adding js aggregator to AMP transformer.
