+++ content/src/content/jcr_root/apps/core/wcm/components/carousel/v1/carousel/customheadlibs.amp.html
@@ -1,0 +1,1 @@
+<script async custom-element="amp-carousel" src="https://cdn.ampproject.org/v0/amp-carousel-0.2.js"></script>
